backend/requirements.txt
Package                       Version
----------------------------- ---------
asgiref                       3.5.2
certifi                       2022.12.7
cffi                          1.15.1
charset-normalizer            2.1.1
coreapi                       2.3.3
coreschema                    0.0.4
cryptography                  38.0.4
defusedxml                    0.7.1
Django                        4.1.4
django-filter                 22.1
django-templated-mail         1.1.1
djangorestframework           3.14.0
djangorestframework-simplejwt 4.8.0
djoser                        2.1.0
idna                          3.4
itypes                        1.2.0
Jinja2                        3.1.2
MarkupSafe                    2.1.1
oauthlib                      3.2.2
Pillow                        9.3.0
pip                           22.3.1
psycopg2-binary               2.9.5
pycparser                     2.21
PyJWT                         2.6.0
python-dotenv                 0.21.0
python3-openid                3.2.0
pytz                          2022.6
requests                      2.28.1
requests-oauthlib             1.3.1
setuptools                    58.1.0
six                           1.16.0
social-auth-app-django        4.0.0
social-auth-core              4.3.0
sqlparse                      0.4.3
uritemplate                   4.1.1
urllib3                       1.26.13
